fixed polling backend to solve crashing issue
the polling backend crashed on ocassions where
the backend attempted to retrieve an incoming
message that had been processed for a particular
sender. The crash occured when there were more
one object - signifying several unprocessed
incoming messages from the same person. This is
a temporary fix to fetch the first of such messages
a more elegant solution will be to use a message
id instead of the identity and message status in
select the message.