added a Campaign Related Objects method
added a Campaign Related Objects method to the campaign model
to help in fetching objects that are used in the campaign. e.g.
Reports, Shortages, ...

The usage is campaign_obj.cro(Object, state, locations)

Object must contain the following fields:
location - This is the location where the data was capatured
time - This is where the time the data was captured.

Any object that fulfills this criteria can be used with the method.

The utility of a method like this is that you can quickly filter
data for a particular campaign using this method.

E.g. give me all the Shortage reports in a particular campaign within
a particular location.